Refactor to make actor more integrated
Move a lot of state into the actor.
This streamlines the interface between the actor and the rest of the code.
Additionally, it reduces the amount of commands and responses that need to move to/from the actor.

I also reimplemented the random shuffle functionality to be deterministic and avoid repetition.