Adjust arrow-key navigation behavior
First of all, add P and N as synonyms for the arrow keys, with Shift-N as a synonym for P.
Next, make all these keys respect the slideshow behavior, meaning that if we are in a random slideshow, the keys will move the same as the slideshow. We allow Alt as a modifier to ignore the slideshow and move normally.